app/requirements.txt
boto3==1.7.2
botocore==1.10.2
celery==4.2.1
Django==2.1.3
django-cors-headers==2.2.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
django-prometheus==1.0.11
django-storages==1.6.6
psycopg2==2.7.3.2
python-logstash==0.4.6
python-dotenv==0.7.1
raven==6.8.0
django-cleanup==2.1.0
PyGithub==1.43.3
redis==2.10.6